pyproject.toml
[project]
name = "pc-agent-downloads-cleanup"
version = "0.1.0"
description = "LangGraph agent to clean up Windows Downloads with human-in-the-loop approvals."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
  "langgraph>=0.2.0",
]

[project.scripts]
pc-agent = "pc_agent.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
